rename: CodeMaker 클래스의 이름 변경 및 패키지 이동
- 직관적인 이름으로 클래스 이름을 변경
- 기존의 util에서 더 적절한 패키지 위치로 이동
